requirements.txt
numpy
opencv-python >= 3.0.0
PyOpenGL
glfw
pillow
pyfiglet
pygame
picamera2